Move React User Components into their own Folder
